+++ shardingsphere-jdbc-spring/shardingsphere-jdbc-transaction-spring/src/main/java/org/apache/shardingsphere/driver/spring/transaction/ShardingTransactionTypeScanner.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.transaction.spring;
+package org.apache.shardingsphere.driver.spring.transaction;
 
 import org.springframework.aop.framework.AbstractAdvisingBeanPostProcessor;
